% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_report.r
\name{na_report}
\alias{na_report}
\title{Report Missing Values}
\usage{
na_report(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame of class "na_report" with columns variable, missing, percent
}
\description{
Returns counts and percentages of NAs for each column in a data frame.
}
\examples{
na_report(airquality)
}
