% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_report.r
\name{print.na_report}
\alias{print.na_report}
\title{Print method for na_report}
\usage{
\method{print}{na_report}(x, ...)
}
\arguments{
\item{x}{An na_report object}

\item{...}{Unused}
}
\description{
Print method for na_report
}
