% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_report.r
\name{na_report}
\alias{na_report}
\title{Report Missing Values in a Data Frame}
\usage{
na_report(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
An object of class "na_report" with NA counts and percentages
}
\description{
Counts NA values for each column and returns a summary.
}
\examples{
na_report(airquality)
}
